fix(ns-ha): correctly retrieve the current state
Previously, the state was read from /tmp/keepalived.stats file but it
was wrong: the stats file returns the name of the vrrp instance, not
its current state.

The new implementation reads the state directly from the log file.
Please note that the state may not be present on a long running
firewall.